test: move remaining live Tenderly tests behind integration tag
Claude #727: three Tenderly/Sepolia tests and BenchmarkTenderlySimulation still ran in unit CI because they were left in tenderly_client_test.go. Move them to tenderly_client_live_integration_test.go so go test without -tags=integration never hits live Tenderly.
